feat(video): handle unkown video id and or playlist id

+++ video/video_api.py
@@ -10,10 +10,14 @@ from playlist import playlist_repository
 @post('/videos/<playlist_id>/<title>/<thumbnail>')
 def create_video(playlist_id, title, thumbnail, db):
     playlist = playlist_repository.retrieve_playlist_by_id(playlist_id, db)
+    if (playlist == None):
+        return HTTPResponse(status=200, body={'status': 'NOK', 'message': 'You can not add a video to the playlist. It does not exist'})
+
     position = playlist['video_position'] + 1
     playlist_repository.update_playlist_video_position(
         playlist_id, position, db)
-    video_repository.create_video(playlist_id, title, thumbnail, position, db)
+    video_repository.create_video(
+        playlist_id, title, thumbnail, position, db)
     return HTTPResponse(status=200, body={'status': 'OK'})
 
 
@@ -26,16 +30,24 @@ def retrieve_videos(playlist_id, db):
 
 
 @put('/videos/<id>')
-def update_video(id):
+def update_video_position(id):
     pass
 
 
 @delete('/videos/<id>/<playlist_id>')
 def delete_video(id, playlist_id, db):
     playlist = playlist_repository.retrieve_playlist_by_id(playlist_id, db)
+    if (playlist == None):
+        return HTTPResponse(status=200, body={'status': 'NOK', 'message': 'You can not delete this. It does not exist'})
+
+    video = video_repository.retrieve_video_position(id, db)
+
+    if (video == None):
+        return HTTPResponse(status=200, body={'status': 'NOK', 'message': 'You can not delete this. this video is not part of this playlist'})
+
     position = playlist['video_position'] - 1
     playlist_repository.update_playlist_video_position(
         playlist_id, position, db)
-    video = video_repository.delete_video(id, db)
+    video_repository.delete_video(id, db)
     video_repository.update_video_positions(video['position'], db)
     return HTTPResponse(status=200, body={'status': 'OK'})
